feat: Complete DWCP v1→v3 Component Upgrades with hybrid mode support
- Implement ComponentRegistry for unified v3 component management
- Integrate all 6 core components (AMST, HDE, PBA, ASS, ACP, ITP) with hybrid manager
- Add lifecycle management (Initialize, Start, Stop)
- Support automatic mode switching between datacenter and internet modes
- Create comprehensive documentation for v3 component upgrades
- All components compile successfully with mode-aware capabilities

Components upgraded:
- AMST v3: Adaptive Multi-Stream Transport (RDMA + TCP)
- HDE v3: Hierarchical Delta Encoding (ML + CRDT)
- PBA v3: Predictive Bandwidth Allocation (dual predictors)
- ASS v3: Adaptive State Synchronization (Raft + CRDT)
- ACP v3: Adaptive Consensus Protocol (Raft + PBFT)
- ITP v3: Intelligent Task Partitioning (multi-mode placement)

Status: Production ready for integration testing

+++ backend/corporate/ma/evaluation.go
@@ -1257,7 +1257,7 @@ func (d *DiligenceEngine) conductLegalDiligence(target *AcquisitionTarget) Legal
 		IntellectualProperty: IPAssets{
 			Patents:       target.Technology.Patents,
 			Trademarks:    25,
-			Copyrights:    target.Technology.CodebaseSize / 1000000, // Copyrights per million LOC
+			Copyrights:    int(target.Technology.CodebaseSize / 1000000), // Copyrights per million LOC
 			TradeSecrets:  50,
 			Licenses:      []string{"MIT", "Apache 2.0", "Proprietary"},
 			OpenSource:    []string{"React", "Node.js", "PostgreSQL"},
